+++ connectors/postgresql.py
@@ -1,4 +1,4 @@
-from .dbms_connector import BasicConnector
+from .basic_connector import BasicConnector
 import psycopg2
 import psycopg2.extras
 import psycopg2.pool
